Update agent_gazebo.py
Updating actor to use Beta Distribution for the policy distribution. The Actor Network now outputs alpha and beta, which define the beta distribution. This distribution is then sampled, and scaled to the range of the action range (either position or effort control). Position control limits are currently set, and effort limits will be set shortly.